# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "API do controle de homologação de projetos solares da ConectSol.\n\nAutenticação: POST /api/auth/login (e-mail e senha). Use o accessToken como Bearer. Não há auto-cadastro: as contas são criadas por ADMINISTRADOR ou GESTOR em POST /api/usuarios. Não há logout no servidor — a API é stateless e o cliente descarta os tokens.\n\nMudanças de status são feitas por endpoints de ação (ex.: POST /api/pendencias/{id}/resolver), nunca por PUT: é o que garante a automação entre etapas e a auditoria.",
    "title" : "SolarSync API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/clientes/{id}" : {
      "delete" : {
        "operationId" : "excluir_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "Clientes" ]
      },
      "get" : {
        "operationId" : "buscar_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClienteResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Clientes" ]
      },
      "put" : {
        "operationId" : "atualizar_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ClienteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClienteResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Clientes" ]
      }
    },
    "/api/clientes/{id}/historico" : {
      "get" : {
        "operationId" : "historico_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/HistoricoStatusResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Linha do tempo da triagem de pendência do cliente",
        "tags" : [ "Clientes" ]
      }
    },
...
    "/api/debitos/{id}" : {
      "delete" : {
        "description" : "Restrito a ADMINISTRADOR. O caminho normal é registrar QUITADO, que preserva o histórico.",
        "operationId" : "excluir_6",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Exclui o registro de débito",
        "tags" : [ "Débitos" ]
      }
    },
    "/api/debitos/{id}/historico" : {
      "get" : {
        "description" : "É desta linha do tempo que sai a métrica de tempo parado por débito.",
        "operationId" : "historico_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/HistoricoStatusResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Linha do tempo do débito",
        "tags" : [ "Débitos" ]
      }
    },
...
    "/api/pendencias/{id}" : {
      "delete" : {
        "description" : "Restrito a ADMINISTRADOR. O caminho normal para registro errado é cancelar por status, que preserva o histórico.",
        "operationId" : "excluir_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Exclui a pendência",
        "tags" : [ "Pendências" ]
      },
      "get" : {
        "operationId" : "buscar_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PendenciaResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Pendências" ]
      },
      "put" : {
        "description" : "Não altera status: use os endpoints de ação.",
        "operationId" : "atualizar_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PendenciaAtualizarRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PendenciaResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Edita os dados da pendência",
        "tags" : [ "Pendências" ]
      }
    },
...
    "/api/pendencias/{id}/historico" : {
      "get" : {
        "operationId" : "historico_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/HistoricoStatusResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Linha do tempo das mudanças de status da pendência",
        "tags" : [ "Pendências" ]
      }
    },
...
    "/api/projetos" : {
      "get" : {
        "operationId" : "listar_3",
        "parameters" : [ {
          "in" : "query",
          "name" : "filtro",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/ProjetoFiltro"
          }
        }, {
          "in" : "query",
          "name" : "paginacao",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaginaResponseProjetoResumoResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Lista projetos paginados, com filtros combináveis",
        "tags" : [ "Projetos" ]
      },
      "post" : {
        "description" : "Para o caso em que o projeto não veio da resolução de uma pendência.",
        "operationId" : "criar_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjetoCriarRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjetoResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Cria um projeto em RECEBIDO",
        "tags" : [ "Projetos" ]
      }
    },
    "/api/projetos/{id}" : {
      "delete" : {
        "description" : "Restrito a ADMINISTRADOR.",
        "operationId" : "excluir_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Exclui o projeto",
        "tags" : [ "Projetos" ]
      },
      "get" : {
        "operationId" : "buscar_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjetoResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Projetos" ]
      },
      "put" : {
        "description" : "Não altera status: use os endpoints de ação.",
        "operationId" : "atualizar_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjetoAtualizarRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjetoResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Edita os dados do projeto",
        "tags" : [ "Projetos" ]
      }
    },
    "/api/projetos/{id}/aguardar-envio" : {
      "post" : {
        "description" : "Projeto pronto mas com o envio bloqueado, tipicamente por débito do cliente.",
        "operationId" : "aguardarEnvio",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjetoResponse"
                }
              }
            },
            "description" : "Transição inválida a partir do status atual"
          }
        },
        "summary" : "Marca o projeto como AGUARDANDO_ENVIO",
        "tags" : [ "Projetos" ]
      }
    },
    "/api/projetos/{id}/aprovar" : {
      "post" : {
        "operationId" : "aprovar_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjetoAprovarRequest"
              }
            }
          }
        },
        "responses" : {
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjetoResponse"
                }
              }
            },
            "description" : "Transição inválida — por exemplo aprovar projeto que não foi enviado"
          }
        },
        "summary" : "Registra a aprovação pela Coelba",
        "tags" : [ "Projetos" ]
      }
    },
    "/api/projetos/{id}/corrigir-status" : {
      "post" : {
        "description" : "Restrito a ADMINISTRADOR e com justificativa obrigatória. A auditoria é mantida. Existe para a importação da planilha, em que os dados chegam em qualquer ordem, e para destravar erro operacional.",
        "operationId" : "corrigirStatus",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CorrigirStatusRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjetoResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Corrige o status pulando a máquina de estados",
        "tags" : [ "Projetos" ]
      }
    },
    "/api/projetos/{id}/encaminhar" : {
      "post" : {
        "description" : "Sem dataEncaminhado no corpo, assume hoje. Essa data alimenta a métrica de tempo entre recebimento e envio. Informe o numeroSolicitacao devolvido pela Coelba: é por ele que o retorno por e-mail será casado com o projeto. Exige consulta de débito de HOMOLOGACAO registrada.",
        "operationId" : "encaminhar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjetoEncaminharRequest"
              }
            }
          }
        },
        "responses" : {
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjetoResponse"
                }
              }
            },
            "description" : "Transição inválida, cliente com débito, ou débito não consultado"
          }
        },
        "summary" : "Encaminha o projeto à Coelba",
        "tags" : [ "Projetos" ]
      }
    },
    "/api/projetos/{id}/historico" : {
      "get" : {
        "operationId" : "historico_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/HistoricoStatusResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Linha do tempo das mudanças de status do projeto",
        "tags" : [ "Projetos" ]
      }
    },
    "/api/projetos/{id}/reencaminhar" : {
      "post" : {
        "description" : "Sem numeroSolicitacao no corpo, mantém o número já registrado.",
        "operationId" : "reencaminhar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjetoEncaminharRequest"
              }
            }
          }
        },
        "responses" : {
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjetoResponse"
                }
              }
            },
            "description" : "Transição inválida, cliente com débito, ou débito não consultado"
          }
        },
        "summary" : "Reenvia o projeto após correção de uma reprova",
        "tags" : [ "Projetos" ]
      }
    },
    "/api/projetos/{id}/registrar-instalacao" : {
      "post" : {
        "description" : "Entrada manual, a partir da informação que chega do campo. Não altera o status (que acompanha a homologação na Coelba) e não exige que o projeto esteja aprovado. É o marco de partida da vistoria: sem esta data, solicitar vistoria é bloqueado. Para achar a fila de trabalho, use GET /api/projetos?instalado=true&semVistoria=true.",
        "operationId" : "registrarInstalacao",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegistrarInstalacaoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjetoResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Registra a data em que a usina foi instalada",
        "tags" : [ "Projetos" ]
      }
    },
    "/api/projetos/{id}/reprovar" : {
      "post" : {
        "operationId" : "reprovar_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MotivoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjetoResponse"
                }
              }
            },
            "description" : "Transição inválida a partir do status atual"
          }
        },
        "summary" : "Registra a reprova; o motivo é obrigatório",
        "tags" : [ "Projetos" ]
      }
    },
...
    "/api/unificacoes/{id}" : {
      "delete" : {
        "description" : "Restrito a ADMINISTRADOR.",
        "operationId" : "excluir_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Exclui a unificação",
        "tags" : [ "Unificações" ]
      },
      "get" : {
        "operationId" : "buscar_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UnificacaoResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Unificações" ]
      },
      "put" : {
        "description" : "Não altera os marcos: use as ações de concluir e desligar.",
        "operationId" : "atualizar_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UnificacaoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UnificacaoResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Edita os dados da unificação",
        "tags" : [ "Unificações" ]
      }
    },
...
    "/api/vistorias/{id}" : {
      "delete" : {
        "description" : "Restrito a ADMINISTRADOR.",
        "operationId" : "excluir_5",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Exclui a vistoria",
        "tags" : [ "Vistorias" ]
      },
      "get" : {
        "operationId" : "buscar_5",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VistoriaResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Vistorias" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ClienteResponse" : {
        "type" : "object",
        "properties" : {
          "cidade" : {
            "type" : "string"
          },
          "dataPagamento" : {
            "type" : "string",
            "format" : "date"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "nome" : {
            "type" : "string"
          },
          "statusTriagem" : {
            "type" : "string",
            "enum" : [ "AGUARDANDO_VERIFICACAO", "COM_PENDENCIA", "SEM_PENDENCIA" ]
          },
          "telefone" : {
            "type" : "string"
          },
          "ucCoelba" : {
            "type" : "string"
          },
          "vendedor" : {
            "type" : "string"
          }
        }
      },
...
      "HistoricoStatusResponse" : {
        "type" : "object",
        "properties" : {
          "entidadeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "entidadeTipo" : {
            "type" : "string",
            "enum" : [ "CLIENTE", "PENDENCIA", "DEBITO", "PROJETO", "VISTORIA", "UNIFICACAO" ]
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "ocorridoEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "statusAnterior" : {
            "type" : "string"
          },
          "statusNovo" : {
            "type" : "string"
          },
          "usuarioId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}